+++ spec/nanook_spec.rb
@@ -31,6 +31,16 @@ RSpec.describe Nanook do
     end
   end
 
+  it 'validate_unit! should allow valid units' do
+    expect(Nanook.validate_unit!('raw')).to eq(true)
+    expect(Nanook.validate_unit!(:raw)).to eq(true)
+    expect(Nanook.validate_unit!('nano')).to eq(true)
+  end
+
+  it 'validate_unit! should raise exception for invalid units' do
+    expect { Nanook.validate_unit!('foo') }.to raise_error(Nanook::NanoUnitError)
+  end
+
   it 'should have a block method' do
     expect(Nanook.new.block('some_block')).to be_kind_of(Nanook::Block)
   end